third_party/WebKit/LayoutTests/editing/deleting/delete_line_end_ws.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
    test(() => assert_selection(
        [
            '<div contenteditable style="border: solid red 50px; font-size: 24px; line-height: 48px; padding: 24px; margin: 64px; width: 600px;">',
                'Fourscore and seven years ago our fathers brought forth on |this continent a new nation...',
            '</div>',
        ].join(''),
        'delete',
        [
            '<div contenteditable style="border: solid red 50px; font-size: 24px; line-height: 48px; padding: 24px; margin: 64px; width: 600px;">',
                'Fourscore and seven years ago our fathers brought forth on|this continent a new nation...',
            '</div>',
        ].join('')),
        'Delete a space at end of line.');

    test(() => assert_selection(
        [
            '<div contenteditable style="border: solid red 50px; font-size: 24px; line-height: 48px; padding: 24px; margin: 64px; width: 600px;">',
                'Fourscore and seven years ago our fathers brought forth on         |this continent a new nation...',
            '</div>',
        ].join(''),
        'delete',
        [
            '<div contenteditable style="border: solid red 50px; font-size: 24px; line-height: 48px; padding: 24px; margin: 64px; width: 600px;">',
                'Fourscore and seven years ago our fathers brought forth on|this continent a new nation...',
            '</div>',
        ].join('')),
        'Delete multiple spaces at end of line.');
</script>
